better doc, some light refactoring
